Add a dedicated postings serialization module.
